requirements.txt
# AI Bot Education Dependencies

# Database
SQLAlchemy>=2.0.0
asyncpg>=0.29.0
psycopg2-binary>=2.9.9

# Telegram Bot API (aiogram)
aiogram>=3.3.0

# Web Framework (if needed)
flask>=3.0.0

# Environment Variables
python-dotenv>=1.0.0

# HTTP Requests
requests>=2.31.0

# Logging
loguru>=0.7.0

# Video processing (for getting video dimensions)
opencv-python>=4.8.0